pyproject.toml
[project]
name = "mongodb-rag-agent"
version = "0.1.0"
description = "Agentic RAG system with PostgreSQL + pgvector and Knowledge Graph"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pydantic-ai>=0.1.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "openai>=1.58.0",
    "docling>=2.14.0",
    "docling-core>=2.4.0",
    "transformers>=4.47.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "streamlit>=1.40.0",
    "streamlit-chat>=0.1.0",
    "natasha>=1.6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[project.scripts]
rag-web = "src.web_ui:main"
rag-cli = "src.cli:main"
